backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Authentication & Security
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0

# OAuth
Authlib==1.3.0
requests==2.31.0

# Email
Flask-Mail==0.9.1

# Validation
email-validator==2.1.0

# Utils
python-dateutil==2.8.2